revise gen name
